gcc lint. 	* as.c (main): Move a inside the #if 0 block which uses it. 	* ecoff.c (current_stabs_filename): Make const. 	* frags.h (frag_align_pattern): Declare. 	* gasp.c (new_file): Cast isp to long, and use %ld to print it. 	* config/tc-alpha.h (md_operand): Add cast to void. 	(alpha_do_align): Declare argument types. 	(tc_get_register): Declare. 	(alpha_frob_ecoff_data): Declare. 	* config/tc-alpha.c: Include <ctype.h>. 	(s_mask): Don't declare; does not exist. 	(line_comment_chars): Remove /* from descriptive comment. 	(tc_get_register): Remove unused local reg. 	(tc_gen_reloc): Don't bother to compare unsigned to zero. 	(s_base): Correct warning to actually print register number. 	(md_begin): Remove unused locals retval, lose, and i. 	(alpha_fix_adjustable): Move default case inside switch to avoid 	warning. 	(load_symbol_address): Remove unused locals reloc_addr, p, sym, 	and addend. 	(emit_byte_manip_r): Declare types for all arguments. 	(emit_extract_r, emit_insert_r, emit_mask_r): Likewise. 	(emit_sign_extend, emit_bis_r, s_proc): Likewise. 	(alpha_ip): Use sprint_value to print offsetT value.  Remove 	unused local size.  Remove unused label get_macro. 	(alpha_do_align): Make fill const. 	(md_apply_fix): Remove unused label check_zov.

+++ gas/gasp.c
@@ -3261,7 +3261,7 @@ new_file (name)
     return 0;
 
   if (isp == MAX_INCLUDES)
-    FATAL ((stderr, "Unreasonable include depth (%d).\n", isp));
+    FATAL ((stderr, "Unreasonable include depth (%ld).\n", (long) isp));
 
   sp++;
   sp->handle = newone;
